SoftgLRendercontext.Texture
