+++ SoftGL/RenderContexts/Texture/SoftgLRendercontext.Texture.cs
@@ -5,19 +5,6 @@ using System.Text;
 
 namespace SoftGL
 {
-    struct TextureUnit
-    {
-        public uint texture1D;
-        public uint texture2D;
-        public uint texture2DMultisample;
-        public uint texture2DArray;
-        public uint texture3D;
-        public uint texture2DMultisampleArray;
-        public uint textureCubeMap;
-        public uint textureBuffer;
-        public uint textureRectangle;
-    }
-
     partial class SoftGLRenderContext
     {
         private const int GL_MAX_TEXTURE_IMAGE_UNITS = 8;
