+++ SoftGL/RenderContexts/Utilities/ContextManager.cs
