+++ SoftGL/RenderContexts/Utilities/InternalFormatHelper.cs
