+++ SoftGL/SoftGL.csproj
@@ -126,17 +126,17 @@
     <Compile Include="GLObjects\Framebuffer\Framebuffer.cs" />
     <Compile Include="GLObjects\Framebuffer\Renderbuffer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RenderContexts\ContextManager.cs" />
+    <Compile Include="RenderContexts\Texture\SoftgLRendercontext.Texture.cs" />
+    <Compile Include="RenderContexts\Utilities\ContextManager.cs" />
     <Compile Include="RenderContexts\Framebuffer\SoftGLRenderContext.Framebuffer.cs" />
     <Compile Include="RenderContexts\GLRenderContext.cs" />
     <Compile Include="RenderContexts\GLRenderContext.IDisposable.cs" />
-    <Compile Include="RenderContexts\LastError.cs" />
+    <Compile Include="RenderContexts\Utilities\LastError.cs" />
     <Compile Include="RenderContexts\Framebuffer\SoftGLRenderContext.Renderbuffer.cs" />
     <Compile Include="RenderContexts\SoftGLRenderContext.API.cs" />
     <Compile Include="RenderContexts\SoftGLRenderContext.cs" />
-    <Compile Include="RenderContexts\ContextPair.cs" />
-    <Compile Include="RenderContexts\InternalFormatHelper.cs" />
-    <Compile Include="RenderContexts\TextureUnit.cs" />
+    <Compile Include="RenderContexts\Utilities\InternalFormatHelper.cs" />
+    <Compile Include="RenderContexts\Utilities\TextureUnit.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="GLAPI\GL.Enumerations\DrawMode.png" />
